update lp-pool-example to the latest code

+++ lp-pool-example/scripts/swap.move
@@ -1,8 +1,8 @@
 script {
     use AptosSwap::Scripts;
-    use Sender::Tokens::{BTC, USDT, LP};
+    use Sender::Coins::{BTC, USDT, LP};
 
-    fun swap(acc: signer, pool_addr: address, btc_in: u128, usdt_out_min: u128) {
+    fun swap(acc: signer, pool_addr: address, btc_in: u64, usdt_out_min: u64) {
         Scripts::swap<BTC, USDT, LP>(acc, pool_addr, btc_in, usdt_out_min);
     }
 }
